Move discharge event to Vaccination page
